+++ config/config.go
@@ -3,14 +3,13 @@ package config
 import (
 	"fmt"
 	"math"
-	"path"
+	"time"
 
 	apiv1 "k8s.io/api/core/v1"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/labels"
 
 	wfv1 "github.com/argoproj/argo-workflows/v3/pkg/apis/workflow/v1alpha1"
-	"github.com/argoproj/argo-workflows/v3/workflow/common"
 )
 
 var EmptyConfigFunc = func() interface{} { return &Config{} }
@@ -23,7 +22,7 @@ type ResourceRateLimit struct {
 // Config contain the configuration settings for the workflow controller
 type Config struct {
 
-	// NodeEvents configures how node events are omitted
+	// NodeEvents configures how node events are emitted
 	NodeEvents NodeEvents `json:"nodeEvents,omitempty"`
 
 	// ExecutorImage is the image name of the executor to use when running pods
@@ -59,7 +58,7 @@ type Config struct {
 	KubeletInsecure bool `json:"kubeletInsecure,omitempty"`
 
 	// ArtifactRepository contains the default location of an artifact repository for container artifacts
-	ArtifactRepository ArtifactRepository `json:"artifactRepository,omitempty"`
+	ArtifactRepository wfv1.ArtifactRepository `json:"artifactRepository,omitempty"`
 
 	// Namespace is a label selector filter to limit the controller's watch to a specific namespace
 	// DEPRECATED: support will be remove in a future release
@@ -106,11 +105,15 @@ type Config struct {
 	// PodSpecLogStrategy enables the logging of podspec on controller log.
 	PodSpecLogStrategy PodSpecLogStrategy `json:"podSpecLogStrategy,omitempty"`
 
-	// PodGCGracePeriodSeconds specifies the duration in seconds before the pods in the GC queue get deleted.
-	// Value must be non-negative integer. A zero value indicates that the pods will be deleted immediately
-	// as soon as they arrived in the pod GC queue.
-	// Defaults to 30 seconds.
+	// PodGCGracePeriodSeconds specifies the duration in seconds before a terminating pod is forcefully killed.
+	// Value must be non-negative integer. A zero value indicates that the pod will be forcefully terminated immediately.
+	// Defaults to the Kubernetes default of 30 seconds.
 	PodGCGracePeriodSeconds *int64 `json:"podGCGracePeriodSeconds,omitempty"`
+
+	// PodGCDeleteDelayDuration specifies the duration in seconds before the pods in the GC queue get deleted.
+	// Value must be non-negative integer. A zero value indicates that the pods will be deleted immediately.
+	// Defaults to 5 seconds.
+	PodGCDeleteDelayDuration *metav1.Duration `json:"podGCDeleteDelayDuration,omitempty"`
 
 	// WorkflowRestrictions restricts the controller to executing Workflows that meet certain restrictions
 	WorkflowRestrictions *WorkflowRestrictions `json:"workflowRestrictions,omitempty"`
@@ -144,6 +147,14 @@ func (c Config) GetResourceRateLimit() ResourceRateLimit {
 	}
 }
 
+func (c Config) GetPodGCDeleteDelayDuration() time.Duration {
+	if c.PodGCDeleteDelayDuration == nil {
+		return 5 * time.Second
+	}
+
+	return c.PodGCDeleteDelayDuration.Duration
+}
+
 // PodSpecLogStrategy contains the configuration for logging the pod spec in controller log for debugging purpose
 type PodSpecLogStrategy struct {
 	FailedPod bool `json:"failedPod,omitempty"`
@@ -163,61 +174,6 @@ type KubeConfig struct {
 	VolumeName string `json:"volumeName,omitempty"`
 	// MountPath of the kubeconfig secret, default to '/kube/config'
 	MountPath string `json:"mountPath,omitempty"`
-}
-
-// ArtifactRepository represents an artifact repository in which a controller will store its artifacts
-type ArtifactRepository struct {
-	// ArchiveLogs enables log archiving
-	ArchiveLogs *bool `json:"archiveLogs,omitempty"`
-	// S3 stores artifact in a S3-compliant object store
-	S3 *S3ArtifactRepository `json:"s3,omitempty"`
-	// Artifactory stores artifacts to JFrog Artifactory
-	Artifactory *ArtifactoryArtifactRepository `json:"artifactory,omitempty"`
-	// HDFS stores artifacts in HDFS
-	HDFS *HDFSArtifactRepository `json:"hdfs,omitempty"`
-	// OSS stores artifact in a OSS-compliant object store
-	OSS *OSSArtifactRepository `json:"oss,omitempty"`
-	// GCS stores artifact in a GCS object store
-	GCS *GCSArtifactRepository `json:"gcs,omitempty"`
-}
-
-func (a *ArtifactRepository) IsArchiveLogs() bool {
-	return a != nil && a.ArchiveLogs != nil && *a.ArchiveLogs
-}
-
-type ArtifactRepositoryType interface {
-	IntoArtifactLocation(l *wfv1.ArtifactLocation)
-}
-
-func (a *ArtifactRepository) Get() ArtifactRepositoryType {
-	if a == nil {
-		return nil
-	} else if a.Artifactory != nil {
-		return a.Artifactory
-	} else if a.GCS != nil {
-		return a.GCS
-	} else if a.HDFS != nil {
-		return a.HDFS
-	} else if a.OSS != nil {
-		return a.OSS
-	} else if a.S3 != nil {
-		return a.S3
-	}
-	return nil
-}
-
-// ToArtifactLocation returns the artifact location set with default template key:
-// key = `{{workflow.name}}/{{pod.name}}`
-func (a *ArtifactRepository) ToArtifactLocation() *wfv1.ArtifactLocation {
-	if a == nil {
-		return nil
-	}
-	l := &wfv1.ArtifactLocation{ArchiveLogs: a.ArchiveLogs}
-	v := a.Get()
-	if v != nil {
-		v.IntoArtifactLocation(l)
-	}
-	return l
 }
 
 type PersistConfig struct {
@@ -280,93 +236,6 @@ type PostgreSQLConfig struct {
 type MySQLConfig struct {
 	DatabaseConfig
 	Options map[string]string `json:"options,omitempty"`
-}
-
-// S3ArtifactRepository defines the controller configuration for an S3 artifact repository
-type S3ArtifactRepository struct {
-	wfv1.S3Bucket `json:",inline"`
-
-	// KeyFormat is defines the format of how to store keys. Can reference workflow variables
-	KeyFormat string `json:"keyFormat,omitempty"`
-
-	// KeyPrefix is prefix used as part of the bucket key in which the controller will store artifacts.
-	// DEPRECATED. Use KeyFormat instead
-	KeyPrefix string `json:"keyPrefix,omitempty"`
-}
-
-func (r *S3ArtifactRepository) IntoArtifactLocation(l *wfv1.ArtifactLocation) {
-	k := r.KeyFormat
-	if k == "" {
-		k = path.Join(r.KeyPrefix, common.DefaultArchivePattern)
-	}
-	l.S3 = &wfv1.S3Artifact{S3Bucket: r.S3Bucket, Key: k}
-}
-
-// OSSArtifactRepository defines the controller configuration for an OSS artifact repository
-type OSSArtifactRepository struct {
-	wfv1.OSSBucket `json:",inline"`
-
-	// KeyFormat is defines the format of how to store keys. Can reference workflow variables
-	KeyFormat string `json:"keyFormat,omitempty"`
-}
-
-func (r *OSSArtifactRepository) IntoArtifactLocation(l *wfv1.ArtifactLocation) {
-	k := r.KeyFormat
-	if k == "" {
-		k = common.DefaultArchivePattern
-	}
-	l.OSS = &wfv1.OSSArtifact{OSSBucket: r.OSSBucket, Key: k}
-}
-
-// GCSArtifactRepository defines the controller configuration for a GCS artifact repository
-type GCSArtifactRepository struct {
-	wfv1.GCSBucket `json:",inline"`
-
-	// KeyFormat is defines the format of how to store keys. Can reference workflow variables
-	KeyFormat string `json:"keyFormat,omitempty"`
-}
-
-func (r *GCSArtifactRepository) IntoArtifactLocation(l *wfv1.ArtifactLocation) {
-	k := r.KeyFormat
-	if k == "" {
-		k = common.DefaultArchivePattern
-	}
-	l.GCS = &wfv1.GCSArtifact{GCSBucket: r.GCSBucket, Key: k}
-}
-
-// ArtifactoryArtifactRepository defines the controller configuration for an artifactory artifact repository
-type ArtifactoryArtifactRepository struct {
-	wfv1.ArtifactoryAuth `json:",inline"`
-	// RepoURL is the url for artifactory repo.
-	RepoURL string `json:"repoURL,omitempty"`
-}
-
-func (r *ArtifactoryArtifactRepository) IntoArtifactLocation(l *wfv1.ArtifactLocation) {
-	u := ""
-	if r.RepoURL != "" {
-		u = r.RepoURL + "/"
-	}
-	u = fmt.Sprintf("%s%s", u, common.DefaultArchivePattern)
-	l.Artifactory = &wfv1.ArtifactoryArtifact{ArtifactoryAuth: r.ArtifactoryAuth, URL: u}
-}
-
-// HDFSArtifactRepository defines the controller configuration for an HDFS artifact repository
-type HDFSArtifactRepository struct {
-	wfv1.HDFSConfig `json:",inline"`
-
-	// PathFormat is defines the format of path to store a file. Can reference workflow variables
-	PathFormat string `json:"pathFormat,omitempty"`
-
-	// Force copies a file forcibly even if it exists (default: false)
-	Force bool `json:"force,omitempty"`
-}
-
-func (r *HDFSArtifactRepository) IntoArtifactLocation(l *wfv1.ArtifactLocation) {
-	p := r.PathFormat
-	if p == "" {
-		p = common.DefaultArchivePattern
-	}
-	l.HDFS = &wfv1.HDFSArtifact{HDFSConfig: r.HDFSConfig, Path: p, Force: r.Force}
 }
 
 // MetricsConfig defines a config for a metrics server