+++ test/e2e/cli_test.go
@@ -3,6 +3,7 @@
 package e2e
 
 import (
+	"encoding/json"
 	"fmt"
 	"io/ioutil"
 	"os"
@@ -10,6 +11,8 @@ import (
 	"strings"
 	"testing"
 	"time"
+
+	"sigs.k8s.io/yaml"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/suite"
@@ -383,16 +386,54 @@ func (s *CLISuite) TestRoot() {
 		})
 	})
 	s.Run("List", func() {
-		s.Given().
-			RunCli([]string{"list"}, func(t *testing.T, output string, err error) {
-				if assert.NoError(t, err) {
-					assert.Contains(t, output, "NAME")
-					assert.Contains(t, output, "STATUS")
-					assert.Contains(t, output, "AGE")
-					assert.Contains(t, output, "DURATION")
-					assert.Contains(t, output, "PRIORITY")
-				}
-			})
+		s.Run("DefaultOutput", func() {
+			s.Given().
+				RunCli([]string{"list"}, func(t *testing.T, output string, err error) {
+					if assert.NoError(t, err) {
+						assert.Contains(t, output, "NAME")
+						assert.Contains(t, output, "STATUS")
+						assert.Contains(t, output, "AGE")
+						assert.Contains(t, output, "DURATION")
+						assert.Contains(t, output, "PRIORITY")
+					}
+				})
+		})
+		s.Run("NameOutput", func() {
+			s.Given().
+				RunCli([]string{"list", "-o", "name"}, func(t *testing.T, output string, err error) {
+					if assert.NoError(t, err) {
+						assert.NotContains(t, output, "NAME")
+					}
+				})
+		})
+		s.Run("WideOutput", func() {
+			s.Given().
+				RunCli([]string{"list", "-o", "wide"}, func(t *testing.T, output string, err error) {
+					if assert.NoError(t, err) {
+						assert.Contains(t, output, "PARAMETERS")
+					}
+				})
+		})
+		s.Run("JSONOutput", func() {
+			s.Given().
+				RunCli([]string{"list", "-o", "json"}, func(t *testing.T, output string, err error) {
+					if assert.NoError(t, err) {
+						list := wfv1.Workflows{}
+						assert.NoError(t, json.Unmarshal([]byte(output), &list))
+						assert.Len(t, list, 1)
+					}
+				})
+		})
+		s.Run("YAMLOutput", func() {
+			s.Given().
+				RunCli([]string{"list", "-o", "yaml"}, func(t *testing.T, output string, err error) {
+					if assert.NoError(t, err) {
+						list := wfv1.Workflows{}
+						assert.NoError(t, yaml.UnmarshalStrict([]byte(output), &list))
+						assert.Len(t, list, 1)
+					}
+				})
+		})
 	})
 	s.Run("Get", func() {
 		s.Given().RunCli([]string{"get", "@latest"}, func(t *testing.T, output string, err error) {
@@ -1120,6 +1161,22 @@ func (s *CLISuite) TestCron() {
 		s.Given().RunCli([]string{"cron", "create", "cron/param.yaml", "-p", "message=\"bar test passed\"", "-l", "workflows.argoproj.io/test=true"}, func(t *testing.T, output string, err error) {
 			if assert.NoError(t, err) {
 				assert.Contains(t, output, "bar test passed")
+			}
+		})
+	})
+
+	s.Run("Create Name Override", func() {
+		s.Given().RunCli([]string{"cron", "create", "cron/basic.yaml", "--name", "basic-cron-wf-overridden-name", "-l", "workflows.argoproj.io/test=true"}, func(t *testing.T, output string, err error) {
+			if assert.NoError(t, err) {
+				assert.Contains(t, strings.Replace(output, " ", "", -1), "Name:basic-cron-wf-overridden-name")
+			}
+		})
+	})
+
+	s.Run("Create GenerateName Override", func() {
+		s.Given().RunCli([]string{"cron", "create", "cron/basic.yaml", "--generate-name", "basic-cron-wf-overridden-generate-name-", "-l", "workflows.argoproj.io/test=true"}, func(t *testing.T, output string, err error) {
+			if assert.NoError(t, err) {
+				assert.Contains(t, strings.Replace(output, " ", "", -1), "Name:basic-cron-wf-overridden-generate-name-")
 			}
 		})
 	})